rename: WeatherInfoRepositoryのプロトコルと実装の命名を変更
