+++ ios-training/Repotitory/WeatherInfoRepository.swift
@@ -2,13 +2,34 @@
 //  WeatherInfoRepository.swift
 //  ios-training
 //
-//  Created by 垣本 桃弥 on 2024/02/13.
+//  Created by 垣本 桃弥 on 2024/02/16.
 //
 
 import Foundation
+import YumemiWeather
 
-/// 天気に関する情報を取得するリポジトリ
-protocol WeatherInfoRepository {
+/// YumemiWeather から天気に関する情報を取得するリポジトリ
+struct WeatherInfoRepository: WeatherInfoRepositoryProtocol {
+    // MARK: Dependencies
+    
+    private let apiEncoder: YumemiWeatherAPIEncoder
+    private let apiDecoder: YumemiWeatherAPIDecoder
+    
+    // MARK: Lifecycle
+    
+    init(apiEncoder: YumemiWeatherAPIEncoder, apiDecoder: YumemiWeatherAPIDecoder) {
+        self.apiEncoder = apiEncoder
+        self.apiDecoder = apiDecoder
+    }
+    
+    // MARK: Internal
+    
     /// 天気に関する情報を取得する
-    func fetch(at area: String, date: Date) throws -> WeatherInfo
+    /// - throws: 取得に失敗した場合は YumemiWeatherError を投げる
+    /// - throws: エンコードやデコードに失敗した場合はそれに対応するエラーを投げる
+    func fetch(at area: String, date: Date) throws -> WeatherInfo {
+        let query = try apiEncoder.encodeQuery(at: area, date: date)
+        let response = try YumemiWeather.fetchWeather(query)
+        return try apiDecoder.decodeResponse(response)
+    }
 }
